popup.jsx converted to popup.tsx
